# auto_generate_video_fold6/docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Auto Video Generation System API
  description: |
    ## Overview
    The Auto Video Generation System is a comprehensive AI-powered platform for creating viral short-form videos automatically. 
    
    This API allows you to:
    - Generate engaging video scripts using AI
    - Create stunning visuals with AI image generation
    - Synthesize natural-sounding voiceovers
    - Analyze trending topics and keywords
    - Schedule and publish content across multiple social platforms
    
    ## Authentication
    All API endpoints require authentication using JWT tokens. Include the token in the Authorization header:
    ```
    Authorization: Bearer YOUR_JWT_TOKEN
    ```
    
    ## Rate Limits
    - Free tier: 60 requests per minute
    - Pro tier: 300 requests per minute
    - Enterprise: Custom limits
    
    ## SDKs and Libraries
    Official SDKs are available for:
    - Python
    - JavaScript/TypeScript
    - React/Next.js
    
  version: "1.0.0"
  contact:
    name: Auto Video Generation Team
    email: support@autovideo.com
    url: https://docs.autovideo.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://autovideo.com/terms
servers:
  - url: https://api.autovideo.com/v1
    description: Production server
  - url: https://staging-api.autovideo.com/v1
    description: Staging server
  - url: http://localhost:8000/api/v1
    description: Development server
security:
  - BearerAuth: []
paths:
...
  /auth/register:
    post:
      tags: [Authentication]
      summary: Register a new user
      description: Create a new user account
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegistration'
            example:
              email: "user@example.com"
              username: "demo_user"
              password: "SecurePassword123"
              full_name: "Demo User"
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
  /auth/login:
    post:
      tags: [Authentication]
      summary: User login
      description: Authenticate user and return JWT token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLogin'
            example:
              email: "user@example.com"
              password: "SecurePassword123"
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/me:
    get:
      tags: [Authentication]
      summary: Get current user profile
      description: Retrieve the authenticated user's profile information
      responses:
        '200':
          description: User profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # AI Text Generation Endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UserRegistration:
      type: object
      required: [email, username, password]
      properties:
        email:
          type: string
          format: email
          example: "user@example.com"
        username:
          type: string
          minLength: 3
          maxLength: 50
          example: "demo_user"
        password:
          type: string
          minLength: 8
          example: "SecurePassword123"
        full_name:
          type: string
          maxLength: 255
          example: "Demo User"
    UserLogin:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthResponse:
      type: object
      properties:
        access_token:
          type: string
          description: JWT access token
        token_type:
          type: string
          example: "bearer"
        expires_in:
          type: integer
          description: Token expiration time in seconds
          example: 1800
        user:
          $ref: '#/components/schemas/UserProfile'
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        username:
          type: string
        full_name:
          type: string
        is_verified:
          type: boolean
        created_at:
          type: string
          format: date-time
    UserProfile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        username:
          type: string
        full_name:
          type: string
        avatar_url:
          type: string
          format: uri
        bio:
          type: string
        website:
          type: string
          format: uri
        location:
          type: string
        is_verified:
          type: boolean
        credits:
          type: object
          properties:
            total:
              type: integer
            used:
              type: integer
            remaining:
              type: integer

    # AI Generation Schemas
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
        timestamp:
          type: string
          format: date-time
    ValidationError:
      type: object
      properties:
        error:
          type: string
          example: "validation_error"
        message:
          type: string
          example: "Invalid input data"
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
  responses:
    BadRequest:
      description: Invalid request data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "unauthorized"
            message: "Authentication token required"
...
    Conflict:
      description: Resource already exists
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...